rename `watermark` module to `state_cleaning`
