+++ src/stream/src/common/table/state_cleaning.rs
